Bring back view and map as props and avoid adding them to store
