+++ src/components/data-global-sidebar/data-global-sidebar-styles.module.scss
@@ -1,15 +1,9 @@
 @import 'styles/ui.module';
 @import 'styles/settings';
-@import 'styles/typography-extends';
-@import 'styles/common-animations.module';
 
-.sidebarContainer {
+.container {
   @extend %verticalScrollbar;
-  @include animationFunction($propertyToAnimate: all);
-  position: absolute;
   height: calc(100vh - #{$sidebar-top-margin});
   width: calc(#{$sidebar-width} + 15px);
-  top: $sidebar-top-margin;
-  left: $site-gutter;
   padding-bottom: $site-gutter;
 }
