+++ src/pages/data-globe/data-globe-styles.module.scss
@@ -1,0 +1,12 @@
+@import 'styles/ui.module';
+@import 'styles/settings';
+@import 'styles/common-animations.module';
+
+.halfEarthLogo {
+  position: absolute;
+  width: $he-project-logo-size;
+  height: $he-project-logo-size;
+  top: $layout-top-margin;
+  left: $site-gutter;
+  z-index: $bring-to-front;
+}
