+++ src/scenes/data-scene/data-scene-styles.module.scss
@@ -1,0 +1,16 @@
+@import 'styles/ui.module';
+@import 'styles/settings';
+@import 'styles/common-animations.module';
+
+$top: calc(#{$sidebar-top-margin} + #{$he-project-logo-size});
+
+.sidebarContainer {
+  @extend %verticalScrollbar;
+  @include animationFunction($propertyToAnimate: all);
+  position: absolute;
+  height: calc(100vh - #{$top});
+  width: calc(#{$sidebar-width} + 15px);
+  top: $top;
+  left: $site-gutter;
+  padding-bottom: $site-gutter;
+}
